feat(mcp): 将 OAuth 2.1 协议层迁移到 dasclaw_mcp::auth (F3.2 phase 2 PR 5b)
## 背景 / 目标

ironclaw 桌面端的 `tools/mcp/auth.rs` 原本是 2 341 行的单一文件，把 OAuth 2.1
协议（PKCE、发现、动态客户端注册、授权 URL 构造、code 换 token、刷新、token
存储）和 ironclaw 专属的"如何接收授权码"（系统浏览器、本地回环回调服务器、
可选的托管 OAuth 代理刷新）混在一起。这块协议层在管理后台 web 服务、CLI 等
未来 host 上也会一字不差地用到。

本 PR 把协议层提到 `crates/dasclaw_mcp/src/auth.rs`，让 ironclaw 只保留它特有
的"交付通路"，给后续 host 留出复用入口。

(Scope H：参数化 `register_client(client_name: &str)`，把"协议层不知道自己叫
什么名字"变成显式参数；ironclaw 通过 `DEFAULT_DCR_CLIENT_NAME` 常量统一传
入。)

## 改动范围

- 新增 `crates/dasclaw_mcp/src/auth.rs` (~2 100 行)
  - 顶部模块文档明确写清：什么留在这、什么留给 host、未来 AI 改协议应该看
    哪里。这是按 user mandate "为未来 ai 读这块代码留下信息" 留的导航。
  - 含：`AccessToken`、`AuthError`、`AuthorizationServerMetadata`、
    `ClientRegistrationRequest/Response`、`ClientCredentials`、
    `PkceChallenge`、`ProtectedResourceMetadata`；
  - 协议函数：`build_well_known_uri`、`canonical_resource_uri`、
    `discover_*`、`register_client`（带 `client_name` 形参）、
    `build_authorization_url`、`exchange_code_for_token`、
    `refresh_access_token_direct`、`get_access_token`、`is_authenticated`、
    `store_*`；
  - 共享 SSRF 守卫 `validate_url_safe` 与 per-(server, user) 刷新锁
    `refresh_lock` 都提升为 `pub`，host shim 直接复用。
- `crates/dasclaw_mcp/Cargo.toml` 升到 edition 2024 (let-chains)，新增
  `dasclaw_runtime / tracing / base64 / chrono / rand / sha2 / urlencoding`
  依赖，`reqwest` 启用 `json`，`tokio` 启用 `sync`。
- `crates/dasclaw_mcp/src/lib.rs` 加 `pub mod auth;` 与精确 re-exports。
- `desktop-client/ironclaw/src/tools/mcp/auth.rs` 从 2 341 行精简到 ~330 行的
  host shim：保留 `authorize_mcp_server`（GUI 编排）、`find_available_port`、
  `wait_for_authorization_callback`、`refresh_access_token`（代理 / 直连
  分发），其余 `pub use dasclaw_mcp::auth::*`。代理刷新分支的"快路径"用
  `match` 取代 `if let Ok`，保留对真实 DB / 解密失败的错误归因（与
  `refresh_access_token_direct` 错误形状对齐）。
- `desktop-client/ironclaw/src/extensions/manager.rs` 唯一调用点改为
  `register_client(reg_endpoint, &redirect_uri, DEFAULT_DCR_CLIENT_NAME)`。

## 非目标 (What's NOT in this PR)

- 不动 admin-backend / CLI 的 OAuth 流程，只把可复用的协议层就位。
- 不引入新的 OAuth 功能（DPoP、device flow 等）。
- 不清理"legacy 连字符密钥名"的兜底读路径——保持现状避免强制重新授权。
- 不调整 SSRF 守卫的 IP 列表（保持与拆分前字节级一致）。

## 验证（命令 + 结果）

```bash
cargo check -p dasclaw --tests
# Finished `dev` profile in 1m 59s

cargo nextest run -p dasclaw_mcp
# Summary [50.717s] 118 tests run: 118 passed, 0 skipped

cargo nextest run -p dasclaw -E 'test(/mcp::auth/)'
# Summary [3.685s] 1 test run: 1 passed, 3936 skipped
# (test_refresh_access_token_uses_proxy_when_configured)

cargo clippy --no-deps -p dasclaw_mcp --all-targets -- -D warnings
# Finished `dev` profile in 6.99s

python3.12 scripts/check_no_panics.py --base feat/f3.2-phase2-pr5a-dasclaw-mcp-config-#641
# OK: No panic-inducing calls in changed production code.

python3.12 scripts/check_no_new_ironclaw_literal.py --base feat/f3.2-phase2-pr5a-dasclaw-mcp-config-#641
# OK: No new .ironclaw / IRONCLAW_BASE_DIR literals introduced.
```

3 层代码审查（quality-audit → simplifier → security）：
- Layer 1 BLOCKER：无；NOTE 已修：代理快路径错误形状回归（已对齐 direct）。
- Layer 2 BLOCKER：无；NOTE 已修：删除零调用的 shim 包装 `get_access_token` /
  `is_server_authenticated`，并把 DCR client_name 常量提到 `dasclaw_mcp::auth`
  与所有 DCR 调用点共享。
- Layer 3 BLOCKER：无；SSRF / PKCE / CSRF / 刷新锁 / fail-closed DNS / 代理
  Bearer / 编码方式与拆分前字节级一致。NOTE 已修：测试用例改用
  `DEFAULT_DCR_CLIENT_NAME` 常量。

Stacked PR：base 不是 `xClaw`，是 `feat/f3.2-phase2-pr5a-dasclaw-mcp-config-#641`
(PR #653)。先看 #653 再看本 PR。

## Sources read

- `AGENTS.md` §2、§3、§4、§5、§6、§7
- `docs/plans/architecture-refactor/adr-112-compatibility-evaluation.md`
- `docs/plans/architecture-refactor/adr-113-hook-engine-unification.md`
- `crates/dasclaw_mcp/src/lib.rs` 模块头
- `desktop-client/ironclaw/src/tools/mcp/auth.rs` 拆分前 (~2341 行) 全文
- `desktop-client/ironclaw/src/cli/oauth_defaults.rs` 接口
- `crates/dasclaw_runtime/src/secrets/mod.rs` 与 `store_in_memory.rs`

Closes #641 (F3.2 phase 2 PR 5b of N).
