+++ crates/dasclaw_mcp/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "dasclaw_mcp"
 version = "0.0.0-w6"
-edition = "2021"
+edition = "2024"
 description = "Model Context Protocol (MCP) client building blocks shared across dasclaw hosts."
 license = "Apache-2.0 OR MIT"
 publish = false
@@ -13,16 +13,35 @@ path = "src/lib.rs"
 # `dasclaw_tool::ToolError` is the application-layer error returned by
 # config helpers (added in F3.2 phase 2 PR 5a, #641).
 dasclaw_tool = { path = "../dasclaw_tool" }
+# `dasclaw_runtime::secrets` provides the `SecretsStore` trait used by the
+# OAuth token-storage helpers in `auth` (added in F3.2 phase 2 PR 5b, #641).
+dasclaw_runtime = { path = "../dasclaw_runtime" }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 thiserror = "1"
-# Async filesystem I/O for `load_mcp_servers_from` / `save_mcp_servers_to`.
-tokio = { version = "1", features = ["fs"] }
+# Async filesystem I/O for `load_mcp_servers_from` / `save_mcp_servers_to`,
+# plus async mutex for the per-(server,user) refresh lock in `auth`.
+tokio = { version = "1", features = ["fs", "sync"] }
 # URL parsing for the localhost / loopback detection in `config::is_localhost_url`.
 url = "2"
-# RFC 9110 header name/value validation in `McpServerConfig::validate`.
-reqwest = { version = "0.12", default-features = false }
+# OAuth discovery / token endpoint HTTP client + RFC 9110 header validation.
+reqwest = { version = "0.12", default-features = false, features = ["json"] }
+# Tracing for OAuth redirect debug logging in `auth`.
+tracing = "0.1"
+# OAuth flow building blocks (PKCE, DCR, token storage) used by `auth`.
+base64 = "0.22"
+chrono = { version = "0.4", features = ["serde"] }
+rand = "0.8"
+sha2 = "0.10"
+urlencoding = "2"
 
 [dev-dependencies]
 tempfile = "3"
-tokio = { version = "1", features = ["fs", "macros", "rt"] }
+tokio = { version = "1", features = ["fs", "macros", "rt", "rt-multi-thread", "sync", "net", "time"] }
+# `auth` tests need an in-memory `SecretsStore` and the runtime's secrecy
+# helpers; rebuild dasclaw_runtime with its test-only `in-memory-secrets` and
+# `libsql` features as needed.
+dasclaw_runtime = { path = "../dasclaw_runtime" }
+secrecy = "0.10"
+# `auth` tests spin up a tiny in-memory HTTP server to record refresh requests.
+axum = "0.8"
